split picker and modal
