+++ src/packages/core/tree/index.ts
@@ -10,7 +10,7 @@ export * from './entity-actions/reload-tree-item-children/index.js';
 export * from './entity-actions/move/index.js';
 export * from './entity-actions/duplicate-to/index.js';
 
-export type { UmbTreePickerModalData, UmbTreePickerModalValue } from './tree-picker/index.js';
-export { UMB_TREE_PICKER_MODAL, UMB_TREE_PICKER_MODAL_ALIAS } from './tree-picker/index.js';
+export type { UmbTreePickerModalData, UmbTreePickerModalValue } from './tree-picker-modal/index.js';
+export { UMB_TREE_PICKER_MODAL, UMB_TREE_PICKER_MODAL_ALIAS } from './tree-picker-modal/index.js';
 
 export * from './types.js';
