+++ src/packages/core/tree/tree-picker-modal/manifests.ts
