+++ src/packages/core/tree/tree-picker-modal/tree-picker-modal.element.ts
@@ -1,6 +1,6 @@
 import type { UmbTreeSelectionConfiguration } from '../types.js';
+import { UmbTreeItemPickerContext } from '../tree-item-picker/index.js';
 import type { UmbTreePickerModalData, UmbTreePickerModalValue } from './tree-picker-modal.token.js';
-import { UmbTreeItemPickerContext } from './tree-picker-modal.context.js';
 import type { PropertyValueMap } from '@umbraco-cms/backoffice/external/lit';
 import { html, customElement, state, ifDefined, nothing } from '@umbraco-cms/backoffice/external/lit';
 import { UMB_WORKSPACE_MODAL, UmbModalBaseElement } from '@umbraco-cms/backoffice/modal';
